What i did ? && Next on:
This is an update of an old project, but notes and Bloc 7.x.x syntax will be added with the inclusion of old steps for in-depth learning.

Did:
A Flutter project was created.
Connection with a new Firebase project was established.
Skeletons of the pages were added.
Foundations of the home page and scrolling feature were laid.
Bloc was added for the scrolling feature.
Models were moved.
Registration screen was introduced (Will be updated in the next stage).
This project will be written using the Bloc architecture.

Next on:

User profile and the required view for the general image.
Registration area.